+++ src/main/java/kitchenpos/application/dto/request/ProductRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.application.dto.request;
 
-import kitchenpos.domain.vo.Price;
+import kitchenpos.common.vo.Price;
 
 public class ProductRequest {
