+++ src/main/java/kitchenpos/common/vo/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.vo;
+package kitchenpos.common.vo;
 
 import java.math.BigDecimal;
 import java.util.Objects;
