+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -11,7 +11,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import kitchenpos.domain.vo.Price;
+import kitchenpos.common.vo.Price;
+import kitchenpos.product.domain.Product;
 
 @Entity
 public class MenuProduct {
