+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,13 +1,12 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.vo.Price;
+import kitchenpos.common.vo.Price;
 
 @Entity
 public class Product {
