+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -14,6 +14,7 @@ import kitchenpos.application.dto.response.MenuResponse;
 import kitchenpos.application.dto.response.OrderResponse;
 import kitchenpos.application.dto.response.ProductResponse;
 import kitchenpos.application.dto.response.TableResponse;
+import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
 import kitchenpos.supports.MenuFixture;
 import kitchenpos.supports.MenuGroupFixture;
