+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -7,7 +7,8 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.application.dto.request.ProductRequest;
 import kitchenpos.application.dto.response.ProductResponse;
-import kitchenpos.domain.vo.Price;
+import kitchenpos.common.vo.Price;
+import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
 import kitchenpos.supports.ProductFixture;
 import org.junit.jupiter.api.DisplayName;
