+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -13,6 +13,7 @@ import kitchenpos.application.dto.response.MenuResponse;
 import kitchenpos.application.dto.response.OrderResponse;
 import kitchenpos.application.dto.response.ProductResponse;
 import kitchenpos.application.dto.response.TableResponse;
+import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
 import kitchenpos.supports.MenuFixture;
 import kitchenpos.supports.MenuGroupFixture;
