+++ src/test/java/kitchenpos/domain/vo/PriceTest.java
@@ -3,6 +3,7 @@ package kitchenpos.domain.vo;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
+import kitchenpos.common.vo.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
